requirements.txt
pandas~=1.5.0
papermill~=2.4.0
black
ipython
ipykernel
PyYAML~=6.0
networkx~=2.8.6
matplotlib~=3.6.0